pyproject.toml
[project]
name = "proposal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]
